Properly populate subscription ID column when sending MMS.
Populate the DB column, and feed the correct value into it for the !MSIM
case. Also clean up the code a little to make the subscription ID value
flow a little more readable.

+++ src/com/android/mms/transaction/TransactionService.java
@@ -899,8 +899,8 @@ public class TransactionService extends Service implements Observer {
     }
 
     protected void endMmsConnectivity() {
-        long subId = SubscriptionManager.getOnDemandDataSubId();
-        endMmsConnectivity(Long.toString(subId));
+        int subId = SubscriptionManager.getOnDemandDataSubId();
+        endMmsConnectivity(Integer.toString(subId));
     }
 
     protected void endMmsConnectivity(String subId) {
